chore: Split deepCasingKeys from casingKeys functions
