+++ src/index.ts
@@ -82,12 +82,6 @@ export {
 export type {
   CamelKeys,
   ConstantKeys,
-  DeepCamelKeys,
-  DeepConstantKeys,
-  DeepDelimiterKeys,
-  DeepKebabKeys,
-  DeepPascalKeys,
-  DeepSnakeKeys,
   DelimiterKeys,
   KebabKeys,
   PascalKeys,
@@ -96,6 +90,22 @@ export type {
 export {
   camelKeys,
   constantKeys,
+  delimiterKeys,
+  kebabKeys,
+  pascalKeys,
+  snakeKeys,
+} from './key-casing'
+
+// DEEP KEY CASING
+export type {
+  DeepCamelKeys,
+  DeepConstantKeys,
+  DeepDelimiterKeys,
+  DeepKebabKeys,
+  DeepPascalKeys,
+  DeepSnakeKeys,
+} from './deep-key-casing'
+export {
   deepCamelKeys,
   deepConstantKeys,
   deepDelimiterKeys,
@@ -103,8 +113,4 @@ export {
   deepPascalKeys,
   deepSnakeKeys,
   deepTransformKeys,
-  delimiterKeys,
-  kebabKeys,
-  pascalKeys,
-  snakeKeys,
-} from './key-casing'
+} from './deep-key-casing'
