+++ src/key-casing.types.test.ts
@@ -1,18 +1,6 @@
 import type * as Subject from './key-casing'
 
 namespace TypeTransforms {
-  type deepTest = Expect<
-    Equal<
-      Subject.DeepDelimiterKeys<
-        {
-          some: { 'deep-nested': { value: true } }
-          'other-value': true
-        },
-        '@'
-      >,
-      { some: { 'deep@nested': { value: true } }; 'other@value': true }
-    >
-  >
   type test = Expect<
     Equal<
       Subject.DelimiterKeys<
@@ -25,15 +13,6 @@ namespace TypeTransforms {
       { 'some@value': { 'nested-value': true }; 'other@value': true }
     >
   >
-  type deepTest1 = Expect<
-    Equal<
-      Subject.DeepCamelKeys<{
-        some: { 'deep-nested': { value: true } }
-        'other-value': true
-      }>,
-      { some: { deepNested: { value: true } }; otherValue: true }
-    >
-  >
   type test1 = Expect<
     Equal<
       Subject.CamelKeys<{
@@ -41,15 +20,6 @@ namespace TypeTransforms {
         'other-value': true
       }>,
       { someValue: { 'deep-nested': true }; otherValue: true }
-    >
-  >
-  type deepTest2 = Expect<
-    Equal<
-      Subject.DeepSnakeKeys<{
-        some: { 'deep-nested': { value: true } }
-        'other-value': true
-      }>,
-      { some: { deep_nested: { value: true } }; other_value: true }
     >
   >
   type test2 = Expect<
@@ -61,15 +31,6 @@ namespace TypeTransforms {
       { some_value: { 'deep-nested': true }; other_value: true }
     >
   >
-  type deepTest3 = Expect<
-    Equal<
-      Subject.DeepKebabKeys<{
-        some: { deepNested: { value: true } }
-        otherValue: true
-      }>,
-      { some: { 'deep-nested': { value: true } }; 'other-value': true }
-    >
-  >
   type test3 = Expect<
     Equal<
       Subject.KebabKeys<{
@@ -79,15 +40,6 @@ namespace TypeTransforms {
       { 'some-value': { deepNested: true }; 'other-value': true }
     >
   >
-  type deepTest4 = Expect<
-    Equal<
-      Subject.DeepPascalKeys<{
-        some: { 'deep-nested': { value: true } }
-        'other-value': true
-      }>,
-      { Some: { DeepNested: { Value: true } }; OtherValue: true }
-    >
-  >
   type test4 = Expect<
     Equal<
       Subject.PascalKeys<{
@@ -95,15 +47,6 @@ namespace TypeTransforms {
         otherValue: true
       }>,
       { SomeValue: { deepNested: true }; OtherValue: true }
-    >
-  >
-  type deepTest5 = Expect<
-    Equal<
-      Subject.DeepConstantKeys<{
-        some: { 'deep-nested': { value: true } }
-        'other-value': true
-      }>,
-      { SOME: { DEEP_NESTED: { VALUE: true } }; OTHER_VALUE: true }
     >
   >
   type test5 = Expect<
